SMLMVis

Documentation for SMLMVis.

SMLMVis.display

SMLMVis.ellipse

SMLMVis.ob_value

SMLMVis.parse_gt

SMLMVis.parse_pt

SMLMVis.scaleim

SMLMVis.display — Method

display(data; predict=nothing, gt=nothing, pixel_size=nothing)

Returns Makie figure with a slider to scroll through different images with groundtruth markers and prediction circles

Arguments

data: The images matrix of size (sz, sz, 1, nframes)

predict: The prediction array

gt: The groundtruth matrix of size (sz, sz, 6, nframes)

pixel_size: The size of a pixel in mms

SMLMVis.ellipse — Method

ellipse(σx, σy[, xc, yc])

Calculates the x and y coordinates to create an ellipse

Arguments

σx: Semi-major/minor axis

σy: Semi-major/minor axis

xc: Center x coordinate

yc: Center y coordinate

SMLMVis.ob_value — Method

ob_value(object, func)

Find value of graph object and run inputted function to create an Observable

Arguments

object: Graph object

func: Function to run with parameter object.value

SMLMVis.parse_gt — Method

parse_gt(framePos, gt)

Used to help display parse groundtruth matrix

Arguments

framePos: The frame position to parse

gt: The groundtruth matrix of size (sz, sz, 6, nframes)

SMLMVis.parse_pt — Method

parse_pt(framePos, pt)

Used to help display parse prediction matrix

Arguments

framePos: The frame position to parse

pt: The prediction array

SMLMVis.scaleim — Method

scaleim(in)

Scale the input from 0 to 1