SMLMVis

Documentation for SMLMVis.

SMLMVis.arr2imgstack — Method

arr2imgstack(arr::Array{<:Real,3})

Convert an array of 3D images to an array of arrays of type N0f8.

Arguments

arr::Array{<:Real,3}: The array of 3D images to convert.

zoom::Int: The zoom factor to apply to each frame. Default is 1.

Returns

imgstack::Array{Array{Gray{N0f8},2}}

This function converts an 3D array to an array of matrices of type N0f8.

SMLMVis.block_resample — Method

block_resample(array::Array, zoom::Int)

Resample an array by repeating each index a specified number of times. This function effectively increases the size of the array by a factor of zoom in each dimension.

Arguments

array::Array: The array to be resampled.

zoom::Int: The factor by which to increase the size of the array.

Returns

Array: The resampled array.

SMLMVis.save_to_mp4 — Method

save_to_mp4(filename::AbstractString, arr::Array{<:Real,3};
    fps::Int=30, crf::Int=23)

Save an array of 3D images to an MP4 video.

Arguments

filename::AbstractString: The name of the output MP4 file.

arr::Array{<:Real,3}: The array of 3D images to save.

fps::Int: The frames per second of the output video. Default is 30.

crf::Int: The constant rate factor of the output video. Default is 23.

zoom::Int: The zoom factor to apply to each frame. Default is 1.

Returns

nothing

This function saves an array of 3D images to an MP4 video using the specified frames per second and constant rate factor. The output video is saved to the specified output file. The function returns nothing.